+++ src/cue/canvascuehandler.h
@@ -17,6 +17,7 @@ namespace PixelMaestro {
 				DrawTriangle,
 				NextFrame,
 				SetCurrentFrameIndex,
+				SetFrameTiming,
 				SetNumFrames,
 				SetOffset,
 				SetScroll
@@ -50,6 +51,7 @@ namespace PixelMaestro {
 			uint8_t* draw_triangle(uint8_t section_num, uint8_t overlay_num, Colors::RGB color, uint16_t point_a_x, uint16_t point_a_y, uint16_t point_b_x, uint16_t point_b_y, uint16_t point_c_x, uint16_t point_c_y, bool fill);
 			uint8_t* next_frame(uint8_t section_num, uint8_t overlay_num);
 			uint8_t* set_current_frame_index(uint8_t section_num, uint8_t overlay_num, uint16_t index);
+			uint8_t* set_frame_timing(uint8_t section_num, uint8_t overlay_num, uint16_t speed);
 			uint8_t* set_num_frames(uint8_t section_num, uint8_t overlay_num, uint16_t num_frames);
 			uint8_t* set_offset(uint8_t section_num, uint8_t overlay_num, int16_t x, int16_t y);
 			uint8_t* set_scroll(uint8_t section_num, uint8_t overlay_num, int16_t x, int16_t y, bool repeat);